ENH: Modernize package setup
Modernize package setup: adopt PEP 518 and partially 631.

Adopt PEP518 to specify minimum build system requirements for the
package.

Partially comply with PEP631:
- Dependencies are now stored in the `pyproject.toml` file.
- The Cythonization is kept in the `setup.py` file.
- Describing the `flake8` configuration in the `pyproject.toml` is not
  supported yet, so it is kept in the `setup.cfg` file.

The testing and packaging is not yet done through `tox` since the
package has no tests at present and `tox` finishes with a failure as a
consequence.

Documentation:
https://www.python.org/dev/peps/pep-0518/
https://www.python.org/dev/peps/pep-0631/